Configure rx_phy_port_pr_map table's default action with D5's updated (#346)
VSI ID.

fxp-net_linux-networking P4 package has a predefined default action
value which always forwards the packets to a fixed vsi_id, which is
presumed to be the one assigned to the D5 interface. As the vsi_id
assignments per inteface is dynamic (which can change per running
instance), a change in D5's vsi_id could potentally break the D5
communication right after IPU init, where the P4 rules are still not
programmed.

To address this issue, we update the rx_phy_port_pr_map table's default
action with the correct vsi_id of D5 interface as part of the port init
configuration procedures.

Signed-off-by: Velayutham, Arunkumar <arunkumar.velayutham@intel.com>