Using mutex, to serialize access to python client p4rtctl (#383)
* Using mutex, to serialize access to python client p4rtctl, when they are concurrent callers
to RunP4rtCtlCommand API in utils.go
Note::p4rtctl python client, uses fixed value(1 for
election-id). So when they are concurrent
clients invoking it, p4 runtime server(infrap4d) throws error.
Add support for inCluster mode, to help distinguish, so
that when waitForInfraP4d invokes->ResolveServiceIp, we
do Lookup(for P4 service name) only when running in cluster,
and in case of non-cluster mode, we set it to default value.